+++ Source/components/position.hpp
@@ -9,15 +9,15 @@ class PositionComponent : public Component
 public:
 	static PositionComponent* Get( const Entity& entity );
 
-	PositionComponent( Entity& owner, float x, float y );
-	PositionComponent( Entity& owner, sf::Vector2f&& position );
+	PositionComponent( Entity& owner, int x, int y );
+	PositionComponent( Entity& owner, sf::Vector2i&& position );
 	~PositionComponent();
 
 	const std::string& GetType() const;
 
-	const sf::Vector2f& GetPosition() const { return m_position;  }
-	sf::Vector2f& GetPosition() { return m_position; }
+	const sf::Vector2i& GetPosition() const { return m_position;  }
+	sf::Vector2i& GetPosition() { return m_position; }
 
 private:
-	sf::Vector2f m_position;
+	sf::Vector2i m_position;
 };
